Move resetting warning toggles logic to mods_panel_controller
