+++ app/controllers/menu_bar_controller.py
@@ -4,20 +4,20 @@ from PySide6.QtCore import QObject, Slot
 from PySide6.QtGui import QAction
 from PySide6.QtWidgets import QApplication, QLineEdit, QPlainTextEdit, QTextEdit
 
+from app.controllers.mods_panel_controller import ModsPanelController
 from app.controllers.settings_controller import SettingsController
 from app.utils.event_bus import EventBus
 from app.utils.generic import open_url_browser
 from app.views.menu_bar import MenuBar
-from app.views.main_content_panel import ModsPanel
 
 
 class MenuBarController(QObject):
-    def __init__(self, view: MenuBar, settings_controller: SettingsController, mods_panel: ModsPanel) -> None:
+    def __init__(self, view: MenuBar, settings_controller: SettingsController, mods_panel_controller: ModsPanelController) -> None:
         super().__init__()
 
         self.menu_bar = view
         self.settings_controller = settings_controller
-        self.mods_panel = mods_panel
+        self.mods_panel_controller = mods_panel_controller
 
         # Application menu
         instance = QApplication.instance()
@@ -105,7 +105,7 @@ class MenuBarController(QObject):
 
         self.menu_bar.rule_editor_action.triggered.connect(EventBus().do_rule_editor)
 
-        self.menu_bar.reset_all_warnings_action.triggered.connect(self._on_menu_bar_reset_warnings_triggered)
+        self.menu_bar.reset_all_warnings_action.triggered.connect(self._on_reset_warnings_triggered)
 
         # Download menu
 
@@ -199,6 +199,9 @@ class MenuBarController(QObject):
         if initialize:
             EventBus().do_activate_current_instance.emit(current_instance)
 
+    def _on_reset_warnings_triggered(self):
+        self.mods_panel_controller.reset_warnings_signal.emit()
+
     @Slot()
     def _on_menu_bar_check_for_updates_on_startup_triggered(self) -> None:
         is_checked = self.menu_bar.check_for_updates_on_startup_action.isChecked()
@@ -228,21 +231,6 @@ class MenuBarController(QObject):
             focused_widget = app_instance.focusWidget()
             if isinstance(focused_widget, (QLineEdit, QTextEdit, QPlainTextEdit)):
                 focused_widget.paste()
-
-    @Slot()
-    def _on_menu_bar_reset_warnings_triggered(self) -> None:
-        """
-        Resets all warning/error toggles.
-        """
-        active_and_inactive_mods = self.mods_panel.collect_mod_list_items()
-        active_mods = active_and_inactive_mods[0]
-        inactive_mods = active_and_inactive_mods[1]
-        for mod_item in active_mods:
-            package_id = mod_item.metadata_manager.internal_local_metadata[mod_item.uuid]["packageid"]
-            mod_item.reset_warning_signal.emit(package_id)
-        for mod_item in inactive_mods:
-            package_id = mod_item.metadata_manager.internal_local_metadata[mod_item.uuid]["packageid"]
-            mod_item.reset_warning_signal.emit(package_id)
 
     @Slot()
     def _on_menu_bar_wiki_triggered(self) -> None:
